feat: 양수가 아닌 id로 특정 방문 기록 삭제를 시도할 때 예외 처리 기능 구현

+++ backend/src/main/java/com/staccato/visit/controller/VisitController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.staccato.exception.InvalidIdException;
 import com.staccato.visit.service.VisitService;
 
 import lombok.RequiredArgsConstructor;
@@ -14,10 +15,15 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/visits")
 @RequiredArgsConstructor
 public class VisitController {
+    private static final int MIN_ID = 1;
+
     private final VisitService visitService;
 
     @DeleteMapping("{visitId}")
     public ResponseEntity<Void> deleteById(@PathVariable Long visitId) {
+        if (visitId < MIN_ID) {
+            throw new InvalidIdException("방문 기록 식별자는 양수로 이루어져야 합니다.");
+        }
         visitService.deleteById(visitId);
         return ResponseEntity.noContent().build();
     }
